Automated Code Cleanup
We all know that the biggest argument in software is how to format code - so much so that we have tools like Laravel Pint to settle such disputes. But we don't want to run Pint manually! What are we, animals?! Instead, let's write a workflow that will automatically clean up our codebase any time we make a change.